Temporary fix: make test conform to error 500
This test was originally written expecting the service error page to
return a HTTP status of 200. I have modified it to instead expect a
status of 500 (Internal Server Error) and adjusted the expect statement
for checking the title matches.

NOTE: This is temporary in that while it works and I think it's a decent
solution, this should be instead implemented by the testers and then
this commit can be rolled back.